requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
